Error checking added to Business Service layer

+++ src/main/java/com/example/doggyHotel/DogService.java
@@ -13,16 +13,39 @@ public class DogService {
         return dogRepository.findAll();
     }
 
-    public Dog addNewDog(Dog dog){
-        return dogRepository.insert(dog);
+    public Dog addNewDog(Dog dog) throws Exception {
+        if(dog.getEmailAddress().contains(".com") && dog.getEmailAddress().contains("@")) {
+            if(dog.getPhoneNumber().contains("08") && dog.getPhoneNumber().length() > 5) {
+                if(dog.getOwnerName().length() > 3 && dog.getOwnerName().length() < 20){
+                    return dogRepository.insert(dog);
+                }
+                else{ throw new Exception("Invalid Owner Name!"); }
+            }
+            else { throw new Exception("Invalid Phone Number!"); }
+        }
+        else{ throw new Exception("Invalid Email Address!"); }
     }
 
-    public void deleteDog(String id){
-        dogRepository.deleteById(id);
+    public void deleteDog(String id) throws Exception {
+        if(!findById(id)){
+            throw new Exception("Dog with this ID does not exist!");
+        }
+        else {
+            dogRepository.deleteById(id);
+        }
     }
 
-    public List<Dog> findByEmail(String email) {
+    public List<Dog> findByEmail(String email) throws Exception {
         List<Dog> dogs = dogRepository.findByEmailAddress(email);
-        return dogs;
+        if(dogs.isEmpty()){
+            throw new Exception("No dogs with this email are in the database.");
+        }
+        else{
+            return dogs;
+        }
+    }
+
+    public boolean findById(String id){
+        return dogRepository.existsById(id);
     }
 }
